поправил везде отображение пустых страниц Добавил страницу админа

+++ src/main/resources/db/migration/V1_1/V1_1_4__create_table_payment_account.sql
@@ -1,5 +1,5 @@
 CREATE TABLE payment_account (
     payment_account_id SERIAL PRIMARY KEY,
     card_number VARCHAR(20) NOT NULL UNIQUE,
-    current_sum DECIMAL(15, 2)
+    current_sum INT
 );
